+++ tests/test_fill.c
@@ -5,7 +5,7 @@
 
 #include "test_common.h"
 
-void test_fill(caterva_ctx_t *ctx, uint8_t ndim, uint64_t *shape_, uint64_t *pshape_, void *value) {
+void test_fill(caterva_ctx_t *ctx, uint8_t ndim, int64_t *shape_, int64_t *pshape_, void *value) {
 
     caterva_dims_t shape = caterva_new_dims(shape_, ndim);
     caterva_dims_t pshape = caterva_new_dims(pshape_, ndim);
@@ -16,7 +16,7 @@ void test_fill(caterva_ctx_t *ctx, uint8_t ndim, uint64_t *shape_, uint64_t *psh
     caterva_fill(src, shape, value);
 
     /* Fill dest array with caterva_array_t data */
-    double *bufdest = (double *) malloc(src->size * src->sc->typesize);
+    double *bufdest = (double *) malloc((size_t)src->size * src->sc->typesize);
     caterva_to_buffer(src, bufdest);
 
     for (int i = 0; i < src->size; ++i) {
@@ -43,8 +43,8 @@ LWTEST_TEARDOWN(fill) {
 
 LWTEST_FIXTURE(fill, 3_dim) {
     const uint8_t ndim = 3;
-    uint64_t shape_[] = {4, 3, 3};
-    uint64_t pshape_[] = {2, 2, 2};
+    int64_t shape_[] = {4, 3, 3};
+    int64_t pshape_[] = {2, 2, 2};
     double value = 13.1265;
     test_fill(data->ctx, ndim, shape_, pshape_, &value);
 }
@@ -52,8 +52,8 @@ LWTEST_FIXTURE(fill, 3_dim) {
 
 LWTEST_FIXTURE(fill, 5_dim) {
     const uint8_t ndim = 5;
-    uint64_t shape_[] = {14, 23, 12, 11, 8};
-    uint64_t pshape_[] = {5, 12, 5, 3, 4};
+    int64_t shape_[] = {14, 23, 12, 11, 8};
+    int64_t pshape_[] = {5, 12, 5, 3, 4};
     double value = 0.005;
     test_fill(data->ctx, ndim, shape_, pshape_, &value);
 }
